requirements.txt
flask
flask_cors
pymongo
bson.objectid